Modify check_ip to consider internal proxied requests as valid
We would like to run webstatscollector using the varnish logs
in Kafka.  Unlike udp2log, this stream does not contain requests
from the nginx SSL and IPv6 proxies.  Instead of throwing out
all varnish requests that have internal RemoteAddresses, we now
consider if the X-Forwarded-For header is set.  If it is, we
(naively) assume that the request is a valid external request
that has been proxied by a WMF internal node.

Change-Id: Id54308de10d95caf2b5acaa1022385ff430c1856